potential udpated

+++ markov_decision_process.hpp
@@ -12,7 +12,7 @@ namespace hexq {
  */
 typedef int Action;
 typedef int State;
-typedef double Reward;
+typedef float Reward;
 /// Cartesian product: States x Actions
 typedef int StateAction;
